requirement.txt
aniso8601==7.0.0
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-allauth==0.44.0
django-extensions==3.1.1
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-knox==4.1.0
djangorestframework==3.12.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
mysql-client==0.0.1
mysqlclient==2.0.3
oauthlib==3.1.0
promise==2.3
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
Rx==1.6.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
